+++ src/math/vec3.rs
@@ -405,10 +405,10 @@ impl DivAssign<f32> for Vec3 {
     }
 }
 
-impl Index<u32> for Vec3 {
+impl Index<usize> for Vec3 {
     type Output = f32;
 
-    fn index(&self, index: u32) -> &<Self as Index<u32>>::Output {
+    fn index(&self, index: usize) -> &<Self as Index<usize>>::Output {
         match index {
             0 => &self.x,
             1 => &self.y,
@@ -418,8 +418,8 @@ impl Index<u32> for Vec3 {
     }
 }
 
-impl IndexMut<u32> for Vec3 {
-    fn index_mut(&mut self, index: u32) -> &mut <Self as Index<u32>>::Output {
+impl IndexMut<usize> for Vec3 {
+    fn index_mut(&mut self, index: usize) -> &mut <Self as Index<usize>>::Output {
         match index {
             0 => &mut self.x,
             1 => &mut self.y,
